Move pattern search to shared utils directory. Hook up pattern search in query block view.
