+++ packages/block-library/src/utils/search-patterns.js
